Implemented using FormStateService in ItemzTypeDetailsComponent.
This will enable users to be able to see Alerts when form data has changed and they try to navigate away to another views. this is important to provide feedback to users and give them opportunity to decide if they would like to save their data or discard it before moving away from ItemzType.